% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-autoplot.R
\name{autoplot.sm_corpus}
\alias{autoplot.sm_corpus}
\title{Autoplot for sm_corpus}
\usage{
\method{autoplot}{sm_corpus}(
  object,
  type = c("production", "landscape", "thematic", "collab", "equity", "top", "lotka",
    "bradford"),
  ...
)
}
\arguments{
\item{object}{An \code{sm_corpus} object.}

\item{type}{Plot type.}

\item{...}{Additional arguments passed to the plot function.}
}
\value{
A \code{ggplot} object.
}
\description{
Automatically generate a suitable plot for an \code{sm_corpus} based on the
requested type.
}
\examples{
corpus <- sm_example_corpus()
ggplot2::autoplot(corpus, type = "production")
}
\seealso{
Other plots: 
\code{\link{sm_palette_qualitative}()},
\code{\link{sm_plot_bradford}()},
\code{\link{sm_plot_citation_network}()},
\code{\link{sm_plot_collab}()},
\code{\link{sm_plot_equity_dashboard}()},
\code{\link{sm_plot_evolution}()},
\code{\link{sm_plot_heaps}()},
\code{\link{sm_plot_landscape}()},
\code{\link{sm_plot_lotka}()},
\code{\link{sm_plot_production}()},
\code{\link{sm_plot_thematic_map}()},
\code{\link{sm_plot_top}()},
\code{\link{sm_scale_color}()},
\code{\link{sm_theme}()}
}
\concept{plots}
