% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus-class.R
\name{is_sm_corpus}
\alias{is_sm_corpus}
\title{Test if an object is an sm_corpus}
\usage{
is_sm_corpus(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{TRUE} if \code{x} is an \code{sm_corpus}, \code{FALSE} otherwise.
}
\description{
Test if an object is an sm_corpus
}
\examples{
corpus <- sm_example_corpus()
is_sm_corpus(corpus)
}
\seealso{
Other corpus: 
\code{\link{as_sm_corpus}()},
\code{\link{sm_bind_corpora}()},
\code{\link{sm_build_corpus}()},
\code{\link{sm_corpus}()},
\code{\link{sm_dedupe}()},
\code{\link{sm_save_corpus}()},
\code{\link{sm_validate}()},
\code{\link{validate_sm_corpus}()}
}
\concept{corpus}
