% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question-class.R
\name{is_sm_question}
\alias{is_sm_question}
\title{Test if an object is an sm_question}
\usage{
is_sm_question(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical.
}
\description{
Test if an object is an sm_question
}
\seealso{
Other question:
\code{\link[=sm_corpus_for_question]{sm_corpus_for_question()}},
\code{\link[=sm_question]{sm_question()}},
\code{\link[=sm_screen_against_question]{sm_screen_against_question()}},
\code{\link[=sm_screen_regex]{sm_screen_regex()}},
\code{\link[=sm_screen_summary]{sm_screen_summary()}}
}
\concept{question}
