% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-clinical-trials.R
\name{sm_field_clinical_trials}
\alias{sm_field_clinical_trials}
\title{Link corpus works to clinical trials}
\usage{
sm_field_clinical_trials(corpus, ..., call = rlang::caller_env())
}
\arguments{
\item{corpus}{An \code{sm_corpus}.}

\item{...}{Additional arguments (currently unused).}

\item{call}{Caller environment.}
}
\value{
A tibble with \code{work_id} and trial-related information.
}
\description{
Identify works in the corpus that are linked to clinical trials
based on PMIDs or DOIs.
}
\examples{
corpus <- sm_example_corpus()
sm_field_clinical_trials(corpus)

}
\seealso{
Other field-helpers: 
\code{\link{sm_field_funder}()},
\code{\link{sm_field_pubmed_mesh}()}
}
\concept{field-helpers}
