% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-funder.R
\name{sm_field_funder}
\alias{sm_field_funder}
\title{Extract funding information}
\usage{
sm_field_funder(
  corpus,
  source = c("crossref", "openalex"),
  call = rlang::caller_env()
)
}
\arguments{
\item{corpus}{An \code{sm_corpus}.}

\item{source}{Data source for funding info.}

\item{call}{Caller environment.}
}
\value{
A tibble of funders with counts.
}
\description{
Summarise funding sources from corpus metadata.
}
\examples{
corpus <- sm_example_corpus()
sm_field_funder(corpus)

}
\seealso{
Other field-helpers:
\code{\link[=sm_field_clinical_trials]{sm_field_clinical_trials()}},
\code{\link[=sm_field_pubmed_mesh]{sm_field_pubmed_mesh()}}
}
\concept{field-helpers}
