% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-prisma.R
\name{sm_import_rayyan}
\alias{sm_import_rayyan}
\title{Import screening decisions from Rayyan}
\usage{
sm_import_rayyan(path)
}
\arguments{
\item{path}{Path to Rayyan CSV export.}
}
\value{
A tibble of screening decisions.
}
\description{
Read Rayyan CSV export and parse screening decisions.
}
\seealso{
Other screening:
\code{\link[=sm_export_covidence]{sm_export_covidence()}},
\code{\link[=sm_export_rayyan]{sm_export_rayyan()}},
\code{\link[=sm_merge_screening_decisions]{sm_merge_screening_decisions()}},
\code{\link[=sm_screen_prisma]{sm_screen_prisma()}}
}
\concept{screening}
