% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-theme.R
\name{sm_palette_qualitative}
\alias{sm_palette_qualitative}
\title{Qualitative viridis palette}
\usage{
sm_palette_qualitative(n, option = "D")
}
\arguments{
\item{n}{Number of colours.}

\item{option}{Viridis palette option.}
}
\value{
A character vector of hex colours.
}
\description{
Qualitative viridis palette
}
\seealso{
Other plots:
\code{\link[=autoplot.sm_corpus]{autoplot.sm_corpus()}},
\code{\link[=sm_plot_bradford]{sm_plot_bradford()}},
\code{\link[=sm_plot_citation_network]{sm_plot_citation_network()}},
\code{\link[=sm_plot_collab]{sm_plot_collab()}},
\code{\link[=sm_plot_equity_dashboard]{sm_plot_equity_dashboard()}},
\code{\link[=sm_plot_evolution]{sm_plot_evolution()}},
\code{\link[=sm_plot_heaps]{sm_plot_heaps()}},
\code{\link[=sm_plot_landscape]{sm_plot_landscape()}},
\code{\link[=sm_plot_lotka]{sm_plot_lotka()}},
\code{\link[=sm_plot_production]{sm_plot_production()}},
\code{\link[=sm_plot_thematic_map]{sm_plot_thematic_map()}},
\code{\link[=sm_plot_top]{sm_plot_top()}},
\code{\link[=sm_scale_color]{sm_scale_color()}},
\code{\link[=sm_theme]{sm_theme()}}
}
\concept{plots}
