% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-trajectory.R
\name{sm_plot_collab_turnover}
\alias{sm_plot_collab_turnover}
\title{Plot collaborator turnover}
\usage{
sm_plot_collab_turnover(traj, dark = FALSE, ...)
}
\arguments{
\item{traj}{An \code{sm_trajectory} object.}

\item{dark}{Logical; dark mode?}

\item{...}{Additional arguments.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise collaborator retention and turnover across career periods.
}
\examples{
\donttest{
corpus <- sm_example_corpus()
traj <- sm_author_trajectory(corpus, author_id = corpus$authors$author_id[1])
sm_plot_collab_turnover(traj)
}

}
\seealso{
Other trajectory: 
\code{\link{is_sm_trajectory}()},
\code{\link{sm_author_trajectory}()},
\code{\link{sm_plot_topic_pivots}()},
\code{\link{sm_plot_trajectory}()},
\code{\link{sm_researcher_profile}()},
\code{\link{sm_trajectory}}
}
\concept{trajectory}
