% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-evolution.R
\name{sm_plot_evolution}
\alias{sm_plot_evolution}
\title{Plot topic evolution over time}
\usage{
sm_plot_evolution(corpus, time_var = "year", dark = FALSE, ...)
}
\arguments{
\item{corpus}{An \code{sm_corpus} with clusters and years.}

\item{time_var}{Column name for time variable. Default \code{"year"}.}

\item{dark}{Logical; dark mode?}

\item{...}{Additional arguments.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise how research topics/clusters evolve across time periods.
}
\seealso{
Other plots: 
\code{\link{autoplot.sm_corpus}()},
\code{\link{sm_palette_qualitative}()},
\code{\link{sm_plot_bradford}()},
\code{\link{sm_plot_citation_network}()},
\code{\link{sm_plot_collab}()},
\code{\link{sm_plot_equity_dashboard}()},
\code{\link{sm_plot_heaps}()},
\code{\link{sm_plot_landscape}()},
\code{\link{sm_plot_lotka}()},
\code{\link{sm_plot_production}()},
\code{\link{sm_plot_thematic_map}()},
\code{\link{sm_plot_top}()},
\code{\link{sm_scale_color}()},
\code{\link{sm_theme}()}
}
\concept{plots}
