% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-trajectory.R
\name{sm_plot_topic_pivots}
\alias{sm_plot_topic_pivots}
\title{Plot topic pivots}
\usage{
sm_plot_topic_pivots(traj, dark = FALSE, ...)
}
\arguments{
\item{traj}{An \code{sm_trajectory} object.}

\item{dark}{Logical; dark mode?}

\item{...}{Additional arguments.}
}
\value{
A \code{ggplot} object.
}
\description{
Vertical timeline showing when an author's research focus shifted.
}
\examples{
\donttest{
corpus <- sm_example_corpus()
traj <- sm_author_trajectory(corpus, author_id = corpus$authors$author_id[1])
sm_plot_topic_pivots(traj)
}

}
\seealso{
Other trajectory:
\code{\link[=is_sm_trajectory]{is_sm_trajectory()}},
\code{\link[=sm_author_trajectory]{sm_author_trajectory()}},
\code{\link[=sm_plot_collab_turnover]{sm_plot_collab_turnover()}},
\code{\link[=sm_plot_trajectory]{sm_plot_trajectory()}},
\code{\link[=sm_researcher_profile]{sm_researcher_profile()}},
\code{\link{sm_trajectory}}
}
\concept{trajectory}
