% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-theme.R
\name{sm_scale_color}
\alias{sm_scale_color}
\alias{sm_scale_fill}
\title{Viridis colour scale for scimapR}
\usage{
sm_scale_color(option = "viridis", discrete = TRUE, ...)

sm_scale_fill(option = "viridis", discrete = TRUE, ...)
}
\arguments{
\item{option}{Viridis palette option. One of \code{"viridis"}, \code{"magma"},
\code{"plasma"}, \code{"cividis"}, \code{"inferno"}, \code{"mako"}, \code{"rocket"}, \code{"turbo"}.}

\item{discrete}{Logical; discrete or continuous scale?}

\item{...}{Additional arguments passed to the viridis scale function.}
}
\value{
A ggplot2 scale object.
}
\description{
Wraps viridis with sensible defaults for scimapR plots.
}
\seealso{
Other plots: 
\code{\link{autoplot.sm_corpus}()},
\code{\link{sm_palette_qualitative}()},
\code{\link{sm_plot_bradford}()},
\code{\link{sm_plot_citation_network}()},
\code{\link{sm_plot_collab}()},
\code{\link{sm_plot_equity_dashboard}()},
\code{\link{sm_plot_evolution}()},
\code{\link{sm_plot_heaps}()},
\code{\link{sm_plot_landscape}()},
\code{\link{sm_plot_lotka}()},
\code{\link{sm_plot_production}()},
\code{\link{sm_plot_thematic_map}()},
\code{\link{sm_plot_top}()},
\code{\link{sm_theme}()}
}
\concept{plots}
