% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus-coerce.R, R/interop-bibliometrix.R
\name{as_sm_corpus}
\alias{as_sm_corpus}
\alias{as_sm_corpus.sm_corpus}
\alias{as_sm_corpus.data.frame}
\alias{as_sm_corpus.list}
\alias{as_sm_corpus.bibliometrixDB}
\title{Coerce objects to sm_corpus}
\usage{
as_sm_corpus(x, ...)

\method{as_sm_corpus}{sm_corpus}(x, ...)

\method{as_sm_corpus}{data.frame}(x, source_label = "data.frame", ...)

\method{as_sm_corpus}{list}(x, ...)

\method{as_sm_corpus}{bibliometrixDB}(x, source_label = "bibliometrix-import", ...)
}
\arguments{
\item{x}{An object to coerce.}

\item{...}{Additional arguments passed to methods.}

\item{source_label}{Label for provenance tracking.}
}
\value{
An \code{sm_corpus} object.
}
\description{
Generic function to convert various data types to an \code{sm_corpus} object.
}
\examples{
corpus <- sm_example_corpus()
as_sm_corpus(corpus)
as_sm_corpus(corpus$works)

}
\seealso{
Other corpus: 
\code{\link{is_sm_corpus}()},
\code{\link{sm_bind_corpora}()},
\code{\link{sm_build_corpus}()},
\code{\link{sm_corpus}()},
\code{\link{sm_dedupe}()},
\code{\link{sm_save_corpus}()},
\code{\link{sm_validate}()},
\code{\link{validate_sm_corpus}()}
}
\concept{corpus}
