% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory-class.R
\name{is_sm_trajectory}
\alias{is_sm_trajectory}
\title{Test if an object is an sm_trajectory}
\usage{
is_sm_trajectory(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical.
}
\description{
Test if an object is an sm_trajectory
}
\examples{
is_sm_trajectory(1)
}
\seealso{
Other trajectory: 
\code{\link{sm_author_trajectory}()},
\code{\link{sm_plot_collab_turnover}()},
\code{\link{sm_plot_topic_pivots}()},
\code{\link{sm_plot_trajectory}()},
\code{\link{sm_researcher_profile}()},
\code{\link{sm_trajectory}}
}
\concept{trajectory}
