% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-corpus.R
\name{sm_bind_corpora}
\alias{sm_bind_corpora}
\title{Bind two corpora together}
\usage{
sm_bind_corpora(corpus1, corpus2)
}
\arguments{
\item{corpus1}{An \code{sm_corpus}.}

\item{corpus2}{An \code{sm_corpus}.}
}
\value{
An \code{sm_corpus} object.
}
\description{
Row-bind all tables of two \code{sm_corpus} objects.
}
\seealso{
Other corpus: 
\code{\link{as_sm_corpus}()},
\code{\link{is_sm_corpus}()},
\code{\link{sm_build_corpus}()},
\code{\link{sm_corpus}()},
\code{\link{sm_dedupe}()},
\code{\link{sm_save_corpus}()},
\code{\link{sm_validate}()},
\code{\link{validate_sm_corpus}()}
}
\concept{corpus}
