% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-corpus.R
\name{sm_dedupe}
\alias{sm_dedupe}
\title{Deduplicate corpus works by DOI}
\usage{
sm_dedupe(corpus, by = "doi", verbose = TRUE)
}
\arguments{
\item{corpus}{An \code{sm_corpus}.}

\item{by}{Column(s) to deduplicate by. Default \code{"doi"}.}

\item{verbose}{Logical; show progress?}
}
\value{
An \code{sm_corpus} object.
}
\description{
Remove duplicate works, preferring the record with more complete data.
}
\seealso{
Other corpus: 
\code{\link{as_sm_corpus}()},
\code{\link{is_sm_corpus}()},
\code{\link{sm_bind_corpora}()},
\code{\link{sm_build_corpus}()},
\code{\link{sm_corpus}()},
\code{\link{sm_save_corpus}()},
\code{\link{sm_validate}()},
\code{\link{validate_sm_corpus}()}
}
\concept{corpus}
