% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-gephi.R
\name{sm_export_gephi}
\alias{sm_export_gephi}
\title{Export network for Gephi (GEXF format)}
\usage{
sm_export_gephi(network, path)
}
\arguments{
\item{network}{A \code{tbl_graph} object.}

\item{path}{Output file path (should end in \code{.gexf}).}
}
\value{
\code{path} invisibly.
}
\description{
Write a tidygraph network to GEXF XML format for use with Gephi.
}
\examples{
\donttest{
corpus <- sm_example_corpus()
net <- sm_network_cocitation(corpus)
sm_export_gephi(net, tempfile(fileext = ".gexf"))
}

}
\seealso{
Other export:
\code{\link[=sm_export_csv]{sm_export_csv()}},
\code{\link[=sm_export_cytoscape]{sm_export_cytoscape()}},
\code{\link[=sm_export_figure]{sm_export_figure()}},
\code{\link[=sm_export_quarto_report]{sm_export_quarto_report()}},
\code{\link[=sm_export_rds]{sm_export_rds()}},
\code{\link[=sm_export_table]{sm_export_table()}},
\code{\link[=sm_export_vosviewer]{sm_export_vosviewer()}},
\code{\link[=sm_export_zip]{sm_export_zip()}}
}
\concept{export}
