% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-prisma.R
\name{sm_export_rayyan}
\alias{sm_export_rayyan}
\title{Export corpus for Rayyan}
\usage{
sm_export_rayyan(corpus, path)
}
\arguments{
\item{corpus}{An \code{sm_corpus}.}

\item{path}{Output file path.}
}
\value{
\code{path} invisibly.
}
\description{
Write corpus works in RIS format for import into Rayyan.
}
\seealso{
Other screening: 
\code{\link{sm_export_covidence}()},
\code{\link{sm_import_rayyan}()},
\code{\link{sm_merge_screening_decisions}()},
\code{\link{sm_screen_prisma}()}
}
\concept{screening}
