% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-gephi.R
\name{sm_export_vosviewer}
\alias{sm_export_vosviewer}
\title{Export network for VOSviewer}
\usage{
sm_export_vosviewer(network, path)
}
\arguments{
\item{network}{A \code{tbl_graph} object.}

\item{path}{Output file path.}
}
\value{
\code{path} invisibly.
}
\description{
Write network to VOSviewer-compatible format.
}
\seealso{
Other export: 
\code{\link{sm_export_csv}()},
\code{\link{sm_export_cytoscape}()},
\code{\link{sm_export_figure}()},
\code{\link{sm_export_gephi}()},
\code{\link{sm_export_quarto_report}()},
\code{\link{sm_export_rds}()},
\code{\link{sm_export_table}()},
\code{\link{sm_export_zip}()}
}
\concept{export}
