% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screening-prisma.R
\name{sm_merge_screening_decisions}
\alias{sm_merge_screening_decisions}
\title{Merge external screening decisions into corpus}
\usage{
sm_merge_screening_decisions(corpus, decisions)
}
\arguments{
\item{corpus}{An \code{sm_corpus}.}

\item{decisions}{A tibble with screening decisions.}
}
\value{
An \code{sm_corpus} with updated screening.
}
\description{
Add external screening decisions (from Rayyan, Covidence, or manual
spreadsheet) into the corpus screening table.
}
\seealso{
Other screening: 
\code{\link{sm_export_covidence}()},
\code{\link{sm_export_rayyan}()},
\code{\link{sm_import_rayyan}()},
\code{\link{sm_screen_prisma}()}
}
\concept{screening}
