% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-citation-net.R
\name{sm_plot_citation_network}
\alias{sm_plot_citation_network}
\title{Plot citation network}
\usage{
sm_plot_citation_network(corpus, top_n = 50L, dark = FALSE, ...)
}
\arguments{
\item{corpus}{An \code{sm_corpus}.}

\item{top_n}{Number of top works to include.}

\item{dark}{Logical; dark mode?}

\item{...}{Additional arguments.}
}
\value{
A \code{ggplot} object.
}
\description{
Visualise the citation network of top-cited works.
}
\examples{
\dontshow{if (requireNamespace("ggraph", quietly = TRUE)) withAutoprint(\{ # examplesIf}
corpus <- sm_example_corpus()
sm_plot_citation_network(corpus)
\dontshow{\}) # examplesIf}
}
\seealso{
Other plots:
\code{\link[=autoplot.sm_corpus]{autoplot.sm_corpus()}},
\code{\link[=sm_palette_qualitative]{sm_palette_qualitative()}},
\code{\link[=sm_plot_bradford]{sm_plot_bradford()}},
\code{\link[=sm_plot_collab]{sm_plot_collab()}},
\code{\link[=sm_plot_equity_dashboard]{sm_plot_equity_dashboard()}},
\code{\link[=sm_plot_evolution]{sm_plot_evolution()}},
\code{\link[=sm_plot_heaps]{sm_plot_heaps()}},
\code{\link[=sm_plot_landscape]{sm_plot_landscape()}},
\code{\link[=sm_plot_lotka]{sm_plot_lotka()}},
\code{\link[=sm_plot_production]{sm_plot_production()}},
\code{\link[=sm_plot_thematic_map]{sm_plot_thematic_map()}},
\code{\link[=sm_plot_top]{sm_plot_top()}},
\code{\link[=sm_scale_color]{sm_scale_color()}},
\code{\link[=sm_theme]{sm_theme()}}
}
\concept{plots}
